Added 3 of Jax's extensions, and 1 from Nifflas; some updates
Thanks guys.
Extensions from Jax are PackX, BitMask, and VOLINFO (NRG753 object). All
are MMF2SDK. I had to modify PackX quite a bit in order for it to
compile.
Extension from Nifflas is the Stochastic Utility. This is an old EDIF
which I converted to the more recent one. It will compile but may not
load in MMF2.