Repaired DarkEDIF comment parser
Removed all the non-working code embedded into the JSON parser, and the
SDK now uses a pre-processor method, where the comments are removed
before they hit the parser.
Also removed the hack code that allowed JSON objects to be referenced by
index as well as name - it now uses the correct variable access method
instead.
The JSON parser was restored to the same as it was before the DarkEDIF
changes.