Start of liblacewing 0.5.4 reimplement
Re-added liblacewing's latest 0.5.4 from master branch, and copied msvc
folder from 0.5.0.

To recompile liblacewing base library, download master, replace files,
and use msvc with VS 2008.
I tried using the make command as per instructions, but it only made a
".a" file, which is a .lib; not a .dll and a .lib together. That can be
found in the Lib folder for what it's worth.

Currently doesn't load in Fusion, 2.0 or 2.5; I've added messageboxes in
General.cpp to figure out why, but no popups are appearing at all.