Bluewing Client now works globally
Used to drop connections between frames due to lack of owning instances
- now features a disable-able timeout thread, which will warn the user
if the connection is left unowned for too long. There's also a new
action that will set the unowned connection to be destroyed.
I moved most of the secondary event queue to GlobalInfo to allow it to
be completely separate from owning Extension classes. In theory, if the
connection is left unowned:
• In multithreaded-mode: the connection should be retained, pings
responded to, and all other messages just queued, waiting for a new
owner.
• In singlethreaded-mode:  the connection will be retained until it
times out naturally, due to a lack of ping responses. Max of 30 seconds
wait, by default non-lacewing settings.
Note it is not closed in either case, so it's not good practice if you
are genuinely trying to destroy the object and the connection. In
addition, a warning message will be shown if you leave a connection
hanging open with no owners for too long (10 seconds currently).
For fastest cleanup, use Disconnect and enable Full Destroy.

Also fixed a bug with Previous Name; for some reason it was referencing
Client tag.

I've done some work on the Bluewing Server, but nothing amazing nor
close to complete (>1600 errors), so I'm leaving that out of the commit
for now.