Big update to Bluewing Client
• Completely overwrote the liblacewing straight from 0.5.4 branch, with
a couple of bugfixes.
• Memory hog improvements
• Events sped up
• Timeout sped up + auto-close on app end
• UDP fixed in a more normal fashion; see liblacewing's pull request
115. Previously was fixed by a side effect of calloc.
• Made latching onto the Extension threadlock less invasive
• Commented out a useless memcpy
• Due to use of multithreaded processing, certain Lacewing files had to
be renamed. "event.c" and "event.cc" would both use "event.obj",
creating access violations.
• Chucked some of the lacewing dependencies out the window, not being
used. We're keeping OpenSSL because while it's not used, it's referred
to in a ctor and we don't want to make changes to the base library code
that may well be overwritten later.