Bluewing Server update, pre-release
There was a big shift in capitialising and the update to 0.5.4 in
liblacewing between the last Bluewing Server version and this one,
so
there's changes everywhere.

I thought I'd release this commit early,
in order to fix the double-lacewing subfolders in the DarkEDIF\Lib path.
I don't know why it got puzzled, but it did.

At the moment Bluewing
Server is compiling, but not linking. There are still quite a bit of
work to do and some functions have not been ported over from their
Client representives - particularly the LacewingCall functions that run
ACE events. It's also worth pointing out the Server should be
multithreaded too, and that being the case, name request etc responses
can't be handled  synchronously as the old server used to do.
(GenerateEvent was suitable for this, but we can't block the main thread
in multithreading. Since a wait for response is natural there's no
programmatic reason multithreading can't be implemented this way.)
I had
to pull out the flash policy struct into a header file in order to
create the C++ variant of  exposing the "on_error" function
pointer.

I'll be working on the Server over next few days so it should
be up and working by end of this week, unless I'm diverted to another
project.