Updated RelayClient to b70
Stopped client sending channel messages when there was no peers on
channel.
Fixed a liblacewing memory leak when a lw_addr variable was used on
stack. This was a chronic problem in the server during UDP, barely
leaked in client though.
Moved IPv6-mapped-IPv4 addresses to a separate function, now in
PhiAddress.cc. Regular liblacewing users can comment that out if they're
so inclined.
Fixed Client condition IDs being slightly off resulting in
OnChannelListLoop not working.
Replaced all for each (MS only) to a more standard loop syntax.
Moved the abhorrent isclosed out from liblacewing to a Fusion-only usage
via PeerCopy/ChannelCopy.
Changed message handlers to return a bool based on whether the rest of
the message should be processed. Only significant for server, but I'm
not publishing the server code yet. I was getting HTTP clients
connecting to the Lacewing server, so there needed to be a way to drop
the rest of the message instantly.
Switched WSADuplicateSocket for detecting if handle was a socket out for
something that didn't break Windows. See Lacewing.h for details on that.
Removed unnecessary client/udp pointers in relayclient.
Fixed project paths using something non-relative.
Switched compiler to v140_xp, i.e. VS 2015 with XP support. Note it's
not a requirement to have XP support though.
Switched while loops to std::algorithm or for loops as appropriate.
Introduced memory cache on loops so there's no worrying about client
messing up variables mid-loop.
Added ChannelListing_ChannelCount expression for client.

NB: Server is still in build 1 on GitHub. On my machine, I have build 4
code, but I'm not publishing source for that until the Server sponsor
approves.