Lacewing Relay Client update to build 75
Fixed a liblacewing crash caused by expecting a list element to exist.
When the list was empty, the attempted remove would crash.
In Bluewing Client, when a client connected, then disconnected, then app
was closed, the closure of the app would run a disconnect. This
disconnect would try to delete already deleted hooks, causing null
pointer violation.
Also added Dependencies to all DarkEDIF JSON apps. As it's used in
DarkEDIF multilanguage, it should be included in all DarkEDIF JSONs.