Lacewing Blue Server release (build 8)
The sponsor who funded server development, has now released the source.

HTML5 was set up, but no implementation of a HTML5 server library yet. A
lot of the HTML5 client was coded, but without a server, there's no way
to test.
Multithreading not implemented. Single-threaded is the only way.
Has French and Portuguese translations, from Kakise and PharaohAnubis.