Minor fix for Ext Menus
Was a small GC leak. Not important, as Ext Menus was only an object meant for quickly reading an extension's data. Only really useful for C++ developers who want to create a new version of an existing extension, and make sure they get their A/C/E IDs matched up.