Changed VSProps to remove unnecessary DLL output
Not sure if it was a VS 2019 Preview update or somehow snuck in between the many tests of the overhaul, but VS was outputting all the runtime DLLs referenced by the MFX builds. To override this, I've changed OutputDir back into IntDir lookalike, and edited MFX and PDB to explicitly output in MFX folder.
A separate change was getting the PDBs to be auto-deleted, so if Debug is built, then Edittime (without PDB), then the PDBs from the Debug build will be gone from MFX folder.