Update FusionSDK.props; fixes full rebuilding on every build
Fixes full rebuilding every build.
The BUILD_DATE in the compilation changed the commandline for C++ compilation every build, and a change to that is seen by VS as a reason to rebuild the entire project. The build date wasn't used in the C++ compiler anyway; only used in the resource compiler, so I've moved the BUILD_DATE there. YEAR is still kept in C++, just in case. If there's demand, I'd add a BUILD_DATE that resolves to the current day.