DarkEdif SDK v4; added ext dev message response to updater
• Added a new type of updater reply for a nice message to ext dev, instead of one that includes all updater log.
Previous ExtUpdateType::Error is now ConnectionError, and will still include the log.
• Fixed some typos, and reformatted titles of message boxes to remove the hyphen.
• Switched update thread spawn and wait to directly invoking the updater function. This may deprecate the atomics, but I'll keep them for now.
I thought the thread spawn was necessary due to potentially slow internet, but it seems this isn't really happening noticeably; the connection exchange is too fast, being at most a few paragraphs of UTF-8.
I imagine there's just not enough users to busy the server, and being it's under 1kB and the server is on fibre, it's not slow enough for a slow connection to cause problems; but if this changes for some reason, I can store datetime of last update check and make sure it's not too frequent.