Major SDK fix (now v5): fixed issue with float expressions
Major due to the subtlety of the bug instead of the frequency.
This bug occurs when an expression is passed parameters (or generates events) which in turn read an expression from the same object, making the return type be changed for the initial expression.
This will notably occur with float expressions, turning them into reading the float variable in memory as integer (normally resulting in a large integer). Equivalent to:
float f = result; returnType = *(int *)&f;
May also affect string expressions.
Integer expressions don't set rHo.Flags, so integer initial expressions shouldn't be affected.