Add Runtime.CopyString to minimise newbie confusion
Don't want to teach the newbies bad habits with string expressions. To clarify: static strings, and things that won't be freed by your ext, can be passed as returns to strings expressions. But in general, it's safer to use Runtime.CopyString; Fusion will allocate it on its heap, and auto-free it if your expression returns a CopyString.