Fixed a minor issue static-analysed in SDK
Issue would cause a crash, but only if extension filename was in a long path (>127 chars long) and in a Unicode build.
Issue was spotted by the new static analysis, which correctly pointed out it was reading too far.

GetModuleFileName is meant to be passed array length in chars, but was being given array length in bytes, and in Unicode builds a char is 2 bytes, so that is incorrect.