Added + edited Bluewing help file in prep for Unicode
LB has given permission for his help files to be put on the repo. They can be opened in HelpNDoc v7.x Personal Edition, for free; his original was HelpNDoc v2, which can be opened in HelpNDoc v3.
These files have been edited, but still need fleshing out and standardisation, and updating for the times, to a good extent (e.g. James is not a Clickteam member, lacewing-project.org is a spam domain).
It also appears some of the bugs reported in the Bugs page were wrong, so they have been mostly amended. I'm also attempting to add more predictability to the help file; e.g. now you know which order channels will be looped in.