Bluewing Client initial port to Android
This is reusing most of the Windows code from build 95, with Windows-specific handles turned to C++ standard.
However, a lot of those specific functionality ports aren't one-to-one matches to C++ standard materials; for example, CRITICAL_SECTION matches nicely with std::recursive_mutex, but there is no "wait for thread to end, with a timeout" akin to WaitForSingleObject().
There is also no "wait for this variable to change", akin to WaitForSingle/MultipleObjects.

I've switched it to polling atomic-bools but it's not the ideal by any means.

Since the Android port has no "EndApp", the closest being something that occurs when any frame ends, I've made a JavaVM shutdown handler instead. Hopefully, it'll do the same job.