Lacewing help now revision 5
I'm going to be more free with updating the revision.
Every edit, even minor typos, when they're released, will cause a revision change in the help file.

In this edit, in Guides > Hosting a server:
• Added clarity to remove internal/external port confusion, as NaitorStudios got stuck on it
• Fixed a text fragment saying "Examples." which was meant for a link that never existed
• Cleaned up the details about ISP and whose ISP provides IPv6-over-4 tech

I also repaired the Lacewing Blue Client > Conditions > Channel > Is Joined to Channel page, which had faulty indenting in the first line.