Lacewing help now revision 6
Fixed "are so aren't" to "and so aren't" in Unicode notes.
Removed some double-space issues, and in areas where spaces are necessary (such as the endianness article), replaced regular space with non-breaking space so future Replace All won't detect them.
Added note for editors so they are aware.

Put the centered images into NoIndent paragraphs so they are actually centered.