XP compatibility issues - rolling back to VS 2019
The support of XP is very inconsistent, and only VS 2019 16.7 and lower are reasonably consistent with it.
The libraries of VS 2022 use Vista+ features, by virtue of the Windows 7 SDK being the one used with XP toolset. I'll add more workarounds and tests.

I've reported it on VS feedback, but they've been investigating it for a while now.
https://developercommunity.visualstudio.com/t/xp-toolset-doesnt-target-xp-properly/1586002#T-N1590342