Disable ARMv7 ASM entirely on mobile
ARMv7 seems to die on Samsung A10s; a particular model of phone that has an ARM64 processor, but runs in 32-bit mode, effectively like ARMv7. Any parameter passing via ASM causes it to crash.

The assembly programmer is still working on new scripts. We have ARM64, but ARMv7 is fighting with use of floats, and the above issue. Might even be compiler bugs, since the GCC that comes with VS is pretty dated.