Update of Bluewing Client to build 98, Server to build 29
Changes since last published build:
Client + Server:
• Added INTERNET permission checks to Android builds.
• Clean up of logging into DarkEdif LOG functions
• UC tagging implemented in public version

Client + Server, in earlier commits:
• Updated help file to revision 11.
• XP compatibility improved.
• Version expression now includes Unicode or ANSI on Windows. For non-Windows, Unicode compatibility is implied.
• Updated to DarkEdif SDK v13, which includes:
» Added event number output in Android/iOS builds.
» Disabled assembly on mobile platforms for all existing ABIs; fixes some rare mobile CPU layouts, like Samsung A10s, which has a 64-bit ARM CPU permanently in 32-bit mode.

Client, in earlier commits:
• Added connection timeout in Android builds, as OS-level timeout was sometimes over a minute.
• Error messages during connecting phase now have the reason.

Server:
• Now by default runs On Connect and other events Relay Server would, although they can be disabled for efficiency by the actions.
• Fixed chance of crash when a client sets name to blank. Only Server crash I've seen in the wild over the last four months.
• On Connect is now triggered by default, matching Relay Server

Server, in earlier commits:
• Error messages improved for selecting channel by name action.
• Fixed file-related error texts

Still some issues with iOS clients during frame switching, and to a lesser extent Android clients. Isolation seems to point to the ref counter, which is hard to pin down.
I'll have to either port Lacewing's refcount to C++ std::atomic, which means any struct it's used in can't be malloc'd and would need to be moved to the C++ new operator, OR I'd have to port all C++-only headers in Lacewing to C/C++ compatible and use _Atomic() for C.
I'm divided on whether to make Lacewing C++ under the hood, and make it have a C external interface, or to make Lacewing C with a C++ interface. It's hard to say no to that nice std::string_view instead of char * and int...