Added Phi Object (build 4)
This is a patron-only ext, was originally written in the DarkEdif MultiTarget SDK - which once it passed proof-of-concept, it was suspended. Eventually, Lacewing Blue was ported to mobile, using the suspended DMT SDK. This by necessity lead to the improvements of GitHub's newer DarkEdif SDK, and DMT's portability, merging together into GitHub a while back.
I'll be porting these DMT SDK exts into GitHub since most Fusion users aren't going to be compiling their own extensions.