DarkEdif SDK: Fixed update checker reporting unfixable error
When server goes down from a local power cut, as happened recently, it sometimes sends no data at all, instead of a socket error. Which means the empty page is presented to the user in the same way a corrupt page would be.
We don't report errors the user can't fix and are expected (including uncommon but expected behaviours like server down), but we do report page corruption because that's unexpected. Errors that aren't reported are considered a no update needed scenario.
This behaviour can be overridden to always report errors via DarkEdif.ini setting ReportAllUpdateCheckErrors = true.