Update of DarkSocket to build 6
Fixed some trailing commas in DarkSocket JSON file.
Trailing commas aren't valid JSON, so I may have to modify the JSON parser to die when it finds them - currently the parser ignores them, but the Data\Runtime JSON minifier produces a corrupt file.

I'm not sure if it was necessary to re-save the example MFA with new ext version, since there are no properties, but why not.