Update of Bluewing Client to b103, Server to b36
Update for iOS fixes, and for Mac compatibility. Due to Mac using iOS code, barely anything needed to be done, other than compilation of OpenSSL for Mac, and edits to the platform in the version string.
Note the b102/b36/SDK v17 fixes for iOS multiple DarkEdif extensions (1b3fef04c49dfb2c225b81f09892ad8afc7b0514), was not implemented until this commit, due to a mistake in SDK v17 release.

Naturally, the OpenSSL libraries had to be rebuilt for Mac. Maybe I'll use the native Mac SSL functions later, but honestly that's borderline cosmetic priority as we're using the latest OpenSSL 3.0.7, and compilation was simple as iOS compilation was well documented.