SafeSurface D3D11 fixes and move out of Template
My plan was to test it in the template, not have it as something that'll be accidentally copied to new projects.

At any rate, safe surface is a neat idea for cross-platform rendering, but might be too Windows-specific even as a wrapper. At very least, we could save the ext dev from Fusion's lack of surface error reporting.