Cleared up Runtime HWA detection
Renamed Runtime.IsHWA(), as it was implied to be whether app was Direct3D or not, and that's not what it does. It instead returns whether the runtime supports Direct3D or not, whether it's using it currently was ignored.

Added Runtime.GetAppDisplayMode() to serve the same sort of purpose.