DarkEdif: RUNDATA removed, EDITDATA const
NO_DEFAULT_CTORS was renamed to NO_DEFAULT_CTORS_OR_DTORS(), to properly reflect contents. It basically makes a C++ struct only initializable and freeable in C-style manual setup; which is used for Fusion memory that comes in C format and is not expected to be created by an ext.

EDITDATA is now passed to Extension in a const way, as it is useless to edit it and undefined to store it, as the memory lifetime is not guaranteed to be perpetual.

RUNDATA is now unused, and is a ghost class not defined anywhere, although its pointer is used on Windows.
* You should not be using RUNDATA to store data anyway in Edif-based SDKs; all the variables normally stored there should be in Extension.
* RUNDATA is only used in Windows; in Android it's in Java-land only accessible with JNI, and in iOS/Mac it's an Obj-C CObject or CRunExtension class, with very different ways of accessing content; unlike EDITDATA, which is stored inside MFA/built apps as binary blocks, and thus is used in all platforms.
* For the default rom/roc etc variables, such as alt values, it's an assumption that RunObject has them. The JNI requiring functions on C++ side is used to check the OEFLAGS and make sure the potential substruct of RunObject eixsts.
* For subwindow or window subclassing, that's Windows only, and you don't actually have to store them in RUNDATA, despite hoOffsetToWindows. The offset is 32-bit int so you can store any 32-bit address, so storing in Extension is fine.
* However, if you're writing a Windows-only ext and want to access RUNDATAs, you can just define it and cast the RunObject passed to Extension ctor to it.

RuntimeFunctions was a plan that allowed cross-platform function calls by using pointers for a unified interface; this turned out unnecessary as Edif Runtime provided it, and there wasn't enough similarity to justify merging them.