DarkEdif SDK v19: Portable object select, Fusion internal vars
Object selection is now DarkEdif namespace, after multiple changes:
• C++ functionality in OIList and similar structs is now implemented.
• Four C++ for-each loop compatible iterators are available, including object info and objects themselves; including qualifiers and standard OI.
• All these work portable on all platforms: Windows (2.0 and 2.5), Android, iOS, Mac.

Internal variables are now get-functions. This is motivated by three things:
• Runtime variables are not in C++ side in Android, but Java-side, requiring a singleton JNI field lookup and JNI function run to read. While this could be pre-loaded in some circumstances, it could become invalid randomly.
• Some variables are moved around per runtime; for example, RunHeader variables are sometimes in CRun, and sometimes in CEventProgram instead.
• Some variables have extra complexity to them; for example, HWA runtimes, differences in RUNDATA structs, or outright missing in some platforms.

However, this does lock down structs to only what is portable and implemented.
To allow Fusion internals to be accessed, for advanced users, you can define FUSION_INTERNAL_ACCESS to disable that locking down.

A new type, RunObjectMultiPlatPtr, is now available; it is a std::shared_ptr on Android, and a vanilla pointer otherwise.
Actions that take Object parameter take a RunObject *, and can choose to instead override Fusion's loop rh4 style.
Instead of
    rhPtr->rh4.ActionStart->evtFlags &= ~ACTFLAGS_REPEAT;
you call
    Runtime.CancelRepeatingObjectAction();
and instead of weirdness with EventParam, you can read an OI with
    Runtime.GetOIFromObjectParam(param_index)
So there is no more ext dev manual pointer arithmetic; it wouldn't work for portability anyway.
For conditions, take an int param, which is actually a short of the possibly qualifier OI entry.

CRunApp is now CRunAppMultiPlat class, as it conflicted with Objective-C CRunApp.

Future upgrades may include copying the runtime selection and restoring it, e.g. to persist selection across generated events. However, it's trivially easy for ext dev to do themselves with the new iterators.

Both PreBuildTool and PostBuildTool are updated to v1.0.0.6, and better support Object and other number-like parameters.