DebugObject: Fix couple crashes
Fixed wrong array size used with reading console input command, and with printing a path of a crashing module.

At some point this is due a rewrite, since Windows console is too limited to work with; a busy output console or user inputting things can cause all sorts of display fights.
It's simpler to just make it a Rich Edit in its own window; but that's for later.