Remove defunct SDKMod idea
SDKMod was an idea of modifying the SDK to create a new ext, and copying ext to Fusion.
Both of these aren't necessary. Prepare DarkEdif tool does a good job with the DarkEdif SDK.
As for copying, you can simply change the Linker Output file path for Windows.

For non-Windows, a new version test is not as simple as copying it to Fusion\Data\Runtime\XX anyway; you have to build an app and deploy it to a device/emulator, so if you automate that, you would automate the copying to Fusion\Data\Runtime to begin with.
And for that non-Windows tester, the WaitForFusionToFinishBuilding app comes in handy. It uses CPU %  tracking to see when Fusion has stopped building, useful for automation.