Bluewing Server: Drop UDP packets if overloaded
If outgoing buffers were full, then UDP packets were creating errors, which while useful for tracking overload, was a terrible idea.
The logging was not efficient, and multiple packets would be denied, so there would be a lot of expensive logging calls, while messages piled up, which quickly filled the outgoing buffer, caused more UDP to get backed, which caused more errors, and it cascaded until the server kicked enough people.
While I don't think this happens in Windows, because Windows uses async overlapped so it approves and queues the messages until it sends them - or maybe it just drops them and says it's finished - it's definitely worth working around in Unix-based OSes (Android, iOS, Mac).