DarkEdif: Allow Android rhEventGroup null
Fixes a JNI crash induced by making global ref of null jobject.
rhEventGroup can be null in Handle(), but it will also be null for rest of an event's actions (and conditions?) if a non-DarkEdif ext has generated an event, as the runtime currently does not save and restore it.