DarkEdif: Look for Android rh4ActionStart
rh4ActionStart is used to get the current action, strangely only Android does not implement it. Windows, iOS/Mac does. Have not looked at HTML5/UWP.

I did some workaround using the CAct when the DarkEdif ext action is used, but outside of that - say Editbox running a DarkScript expression-function - the lack of rh4ActionStart makes it impossible to work with.
The only time it is meant to be null is Handle().

Yves has said he will add it in next runtime and "that's weird" that it doesn't already have it. For those wanting to patch the runtime itself, look at CEventProgram files, and compare iOS runtime and Android runtime. It's only four or so lines to edit.