DarkScript: Add no return value ignore property (b3 start)
This allows you to customise the behaviour when no return value is set. Behaviour before this was equivalent to the "Foreach, delayed only" behaviour: if the expressions were being run in a DarkScript foreach action, or by a delayed DS function run, their return value was not directly accessed, so the lack of a return value was ignored.

I'm still not enthusiastic about how you have to set default return value outside the function template setup action. It doubles the action count of an otherwise pretty smooth 1:1 transition of MFA events from anonymous functions everywhere --> template functions pre-defined.
My current plan is to allow an func() = "XX" after the function declaration, similar to initialised values in C++.

This is start of build 3, due to an extra object property.