DarkScript: Fix running function from script text
This commit has four parts:
1. The parser was incomplete.
It didn't process 2+ parameters or KRFunc type text properly.
Now, the parser can work with basic running of one function in normal format `Bob(1,2)` or in Func format `RFunc("Bob", 5, 1, 2)`.

2. I converted the regexes to use raw-style text. Less escaping is worth the uncommonness.
3. Also added the "running from script" as a part of Sub_GetLocation(), which is used to set up variables used in the call stack expression.
4. Fixed a bug in parameter text parsing where an extra quote, unescaped, would result in two parameters being  merged into one silently.
I lost a couple hours before I realised it wasn't a regex problem, but the value was the problem. GIGO classic.

At some point, the run script from text parsing can become a full script engine, with conditional statements, loops etc.
But I'm a bit wary of that because the Chomsky grammar level of regex is not high enough to parse a context-important script, and writing a full-blown parser is going to be convoluted and require a lot of hours.
As is, just allowing loading one function from script is enough to make loading scripts and Fusion event-based parsers possible.