DarkScript: Fix and protect global func declarations
Global declarations are deceptively simple to use, but rather convoluted in making sure DS functions don't use dead memory of now-defunct extensions.
All exts that use functions in other frames must register their globals so if the DS in that frame is destroyed, the one relying on it doesn't use an invalid Extension pointer.
As further complication, DS allows you to link directly to the template in the other frame, without creating a copy. This means the "fake" GlobalData that was never global, is now queued up for notification - even though it could expire. So now destroying local exts must also check if their globals is used anywhere.

This should cover it; there are likely edge cases, but they should gracefully skip or error out, not crash.