DarkEdif: PostBuildTool custom iOS wrapper fix
PostBuildTool now v1.0.0.8. The iOS wrapper file is h, hpp, and mm files - the h file is linked to automatically by Fusion, and only imports the hpp file.
PostBuildTool looks in the project folder for a custom wrapper file, but for iOS, it would mistakenly look for hpp, hpp, mm, instead of h, hpp, mm in project folder. This new version fixes that.