FusionSDKConfig note multi processor during debug
Add multi processor note to FusionSDKConfig template. It was implemented a while ago.
Multi-processing was normally off by default to allow Debug's Edit And Continue, but that rarely worked in my experience, so in the new ini template it's on by default.

It should be noted that it's always on during non-Debug and non-Windows builds.