DarkScript: Add proof-of-concept HTML5 port
This  HTML5 port includes the new DarkEdif object selection in HTML5 format, which is not in the main DarkEdif Template yet, and may still need work.
I'm kind of wanting to modify Fusion HTML5 runtime; it would make VS Code a lot more happy if the runtime had comments explaining return types and parameter types. Right now, the port instead has to reapply edits for every action and condition, as they aren't implemented in the runtime in a way the prototype can be overridden. See ActionPrototype as evil constant-write, and compare with CObject.prototype.get_Number.
This has yet to be tested in HTML5 Final build, only Development. I expect it will fall over. We'll see.
However, object selection did work in my tests, from memory - but it was better part of a year ago. I'm uploading now because otherwise it'll be months again before it comes to pass - I'm not planning on distributing this version, not until at least Final works.