DarkEdif: Fix SDK config INI path lookup
There was an inconsistency between how the prop file passed the INI path to build tools, and how the props file read it for itself.
Path is now looked up from solution folder, and if not there, from 0 to 2 levels up of project folder, before giving up and using the prop file defaults.