DarkEdif: Fix OIList vs OI num confusion
The OIList number is an index in rhOiList. The Oi number is unique in rhOiList, but is not the index. Oi/OiList is unique per object, but same per instance of that object.

While I think rhOiList is in incremental order, there's a chance it isn't, like included frames in other frames. I've added some aggressive messaging if this is inconsistent.

I have moved OIList parameters in DarkEdif iterators to short. This should make no functional difference.
Other Object Selection parameters are now more clearly defined as shorts. Riggls' framework seems to confuse OI and OIList index, so I may have to correct it - runtimes implement ObjectSelection.cpp in their way, so it may be useful to call into those, in which case the parameter needs to be correct.