DarkEdif: Fix rh2ActionOn during generated events
If rh2ActionOn is false, the runtime is running actions, not conditions. When expressions are false and rh2ActionOn is false, the expressions are assumed to be conditions modifying the selection of objects.
If rh2ActionOn is true, selection is unaffected by expressions. Any mis-match of rh2ActionOn causes selection confusion.