DarkScript: OI used where OIList or hoNum expected
The OIList number is an index in rhOiList. The OI number is unique in rhOiList, but is not the index. Oi/OiList is unique per object, but same per instance of that object.
hoNumber is the object instance number. It is the index in rhObjectList. It is not related to OI or OiList.

This fixes the OI being saved instead of hoNumber for instances, and fixes OI being used in DarkEdif iterators that expect oiList index. Combined with rh2ActionOn fix of commit fae1f13278cdf3272b78ccc6b76084e2351b0fc1 and object selection should now store as expected.