Fix various resource.h not being Resource.h
While it doesn't matter on Windows, when you start cross-building some tools will get upset if file path capitalization is not correct.
I noticed this due to a tool warning, after building on a mirror of the repo.

For those who have these files, you may want to delete Resource.h and download it again.