DarkEdif: Add support for image list properties
All DarkEdif extensions are expected to implement EditProp by passing to DLL_EditProp. See Edittime.cpp edits.
Not implementing it will result in a crash when a button/image list property is pressed.
Currently, DarkEdif ignores standard buttons.

This also includes fixes for mv function names, and the parameter type of EditSurfaceA/W, which was incorrectly an image.

Image-using exts must implement EnumElts function to tell the runtime what images they are using; this is handled for you in DarkEdif.