Bluewing Server: Add support for CG-NAT IPv4 clients
Carrier-Grade NAT often changes the public IP for a client on a per-connection basis. As Bluewing uses TCP and UDP, a client would often have two IP addresses, resulting in UDP packets being discarded as the client they claimed to be had a different IP.
This issue notably happened with IPs starting with 172, *outside* the private 172 range, assigned to T-Mobile. It's worth noting they don't modify IPs for IPv6, and in theory that shouldn't be needed, so this workaround is IPv4 only.

This does present an impersonation risk during the brief connection window, which can only be addressed by removing support for the legacy Relay clients.