Bluewing: Add firewall pinholing for P2P play
Firewall pinholing, to sum it up reductively, is having both server and client connect to each other simultaneously. Firewalls, by default, allow any outgoing connection but not incoming ones. By having an outgoing one exist when an incoming connection comes in, the firewall already has a "hole" to allow the incoming connection to succeed. This requires a fixed local port, for both server and client to know each others IPs and ports, and will only work for 1-to-1 server-client connections.
It's a valid alternative to the IPv4-only uPnP in IPv6 situations. For traditional servers it shouldn't be necessary at all, and since only 1 client is permitted it's not useful any 3-player or bigger game - until I redesign the server to permit a different port and outgoing socket per user, that is.

For servers that will provide gamematching, and allow clients to host their own P2P games, as incoming connections would otherwise be denied in most default firewall configurations, pinholing or use of uPnP is required.

This implementation is sort of rough; I plan to clean it up once I look over my notes on connection timings.