DarkEdif: Set Android 16KB page size for Google Play
As per new Google Play app guidelines (required in apps from Nov 2025), a 16KB page size is required for all libraries, on ARM64, and preferably x86_64. This means you have to rebuild every Android Fusion native extension and every .SO library they link to. The .A libraries require no changes.
https://developer.android.com/guide/practices/page-sizes#compile-16-kb-alignment

16KB is now the default in DarkEdif for ARM64/x86_64.
You can confirm a ELF page size by doing
`readelf -l "path to.so"`
and inspecting the Align column for LOAD sections. Align that is 0x4000 (16384) or a multiple of it is good.