Bluewing: iOS/Mac global namespace conflict fix
The global lacewing namespace in Client and Server will overlap, so for server it has been renamed. This means where their definitions differ - for example, you upgraded Server but not Client, or vice versa - then the iOS app won't be picking one at random.
(It's probably not random, but I'm not sure what Apple Clang prefers to use in these scenarios. First come first served, or overwrite until the last?)