DarkEdif: Only init ext icon if it will load
If image filter manager is not available, the SDK would still allocate memory for a cSurface, but not set it. Now it only constructs if it can load it.
This may cause a crash later if it assumes the Icon ptr is non-null - but I think that might have happened anyway since the cSurface was never Create()'d.