Bluewing Server: fix name set when denied
This is a Fusion server bug, does not affect bluewing-cpp-server.

The nameset_response will send a response message, and also set the client name. For some reason, HandleInteractiveEvent would rename the client after denying it, resulting in a strange scenario where name set would be sent as a deny, but client would be manually renamed on server side anyway.