DarkEdif: Image list fixes
Image titles were borked by last commit, due to reading json_value ptr not ref, so cast to const char * succeeded but was reading the json_value address, not the text within.
An initial minimum/fixed number of images was not possible due to DE always providing only one initially. It is now allowed via NumImages.
The initial image size for editing when no image was there was 1x1, it is now 32x32 initially, if no fixed size is specified.